rename package as jsonencoding
